Adicionando DECIMAL nas migrations e no model

+++ back-end/src/database/migrations/20221027184360-created-products.js
@@ -12,7 +12,7 @@ module.exports = {
         allowNull: false,
       },
       price:{
-        type: Sequelize.FLOAT(10,2),
+        type: Sequelize.DECIMAL(4,2),
         allowNull: false,
       },
       urlImage:{
